+++ docs/source/changelog.rst
@@ -8,6 +8,12 @@ All notable changes to pygacity will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[0.8.0] - 2026-03-11
+--------------------
+
+* updated preamble handling to allow for more flexible formatting
+* more transparent question numbering in the LaTeX output
+
 [0.7.0] - 2026-02-20
 --------------------